requirements.txt
faiss==1.9.0
hnswlib==0.8.0
numpy==2.3.2
Pillow==11.3.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
torchaudio==2.7.1+cu118